extract common
